Fixing bugs in Windows versions of unemap
git-svn-id: https://svn.physiomeproject.org/svn/cmiss/cmgui/trunk@1686 3c1deb5b-d424-0410-962d-aba41a686d42

+++ source/general/matrix_vector.c
@@ -950,7 +950,7 @@ Calculates the inverse of 3X3 FE_value matrix <a>, returning it in <a_inv>.
 
 			for (i = 0; i < 9; i++)
 			{
-				a_inv[i] = d_a_inv[i] / determinant;
+				a_inv[i] = (float)( d_a_inv[i] / determinant);
 			}
 			return_code = 1;
 		}
